+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 <!-- here goes all the unreleased changes descriptions -->
+### Features
+- add support for emitting schema file in not existing directory (#269)
 ### Others
 - **Breaking Change**: drop support for Node.js v6 (end of LTS in April 2019)